requirements.txt
tensorflow
pandas
numpy
scikit-learn
tensorboard
matplotlib
streamlit
scikit-learn
plotly
google-generativeai
python-dotenv